+++ src/rt_scale.c
@@ -414,7 +414,7 @@ void ScaleShape(visobj_t* sprite)
 	}
 
 	// dc_iscale=(1<<(16+6+HEIGHTFRACTION+size))/sprite->viewheight;
-	dc_iscale = 0xffffffffu / (unsigned)dc_invscale;
+	dc_iscale = (int)((double)0xffffffffu / (unsigned)dc_invscale);
 	dc_texturemid =
 		(((sprite->h1 << size) + p->topoffset) << SFRACBITS); //+(SFRACUNIT>>1);
 	sprtopoffset = centeryfrac - FixedMul(dc_texturemid, dc_invscale);
@@ -570,7 +570,7 @@ void ScaleTransparentShape(visobj_t* sprite)
 	}
 
 	//   dc_iscale=(1<<(16+6+HEIGHTFRACTION+size))/sprite->viewheight;
-	dc_iscale = 0xffffffffu / (unsigned)dc_invscale;
+	dc_iscale = (int)((double)0xffffffffu / (unsigned)dc_invscale);
 	dc_texturemid =
 		(((sprite->h1 << size) + p->topoffset) << SFRACBITS); //+(SFRACUNIT>>1);
 	sprtopoffset = centeryfrac - FixedMul(dc_texturemid, dc_invscale);
@@ -885,7 +885,7 @@ void DrawPositionedScaledSprite(int x, int y, int shapenum, int height,
 	if (x2 < 0)
 		return; // off the left side
 
-	dc_iscale = 0xffffffffu / (unsigned)dc_invscale;
+	dc_iscale = (int)((double)0xffffffffu / (unsigned)dc_invscale);
 	//   dc_iscale=(1<<(16+6+size))/height;
 	dc_texturemid =
 		(((32 << size) + p->topoffset) << SFRACBITS) + (SFRACUNIT >> 1);
@@ -973,7 +973,7 @@ void DrawScreenSizedSprite(int lump)
 		return; // off the left side
 	}
 
-	dc_iscale = 0xffffffffu / (unsigned)dc_invscale;
+	dc_iscale = (int)((double)0xffffffffu / (unsigned)dc_invscale);
 	dc_texturemid =
 		(((p->origsize >> 1) + p->topoffset) << SFRACBITS) + (SFRACUNIT >> 1);
 	sprtopoffset =
